unify schema discovery for spark and databricks
